Add heavy attack functionality and inventory improvements
Introduced heavy attack input, animation, and hitbox for the player, including new damage values and movement restrictions during heavy attacks. Refactored damage system to use float values across all damageable entities. Improved inventory menu handling and movement speed adjustments when inventory is open. Updated relevant input bindings, animator controller, and scene objects to support these features.

+++ Assets/Scripts/Enemy/EnemyContactDamage.cs
@@ -27,7 +27,7 @@ public class EnemyContactDamage : MonoBehaviour, IDamageable
         }
     }
 
-    public void TakeDamage(int amount, Vector2 knockbackPosition)
+    public void TakeDamage(float amount, Vector2 knockbackPosition)
     {
         StartCoroutine(HandleDamageFeedback());
     }
