+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,12 +1,6 @@
-// Author: Copilot
-// Version: 1.0
-// Purpose: Interface for objects that can take damage.
-
-
-
 using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int amount, Vector2 position);
+    void TakeDamage(float amount, Vector2 position);
 }
